+++ qsr_prob_rep/src/qsrrep_hmms/hmm_abstractclass.py
